+++ apps/api/package.json
@@ -37,7 +37,7 @@
     "@types/node": "^20",
     "@types/uuid": "^10.0.0",
     "@workspace/api": "workspace:*",
-    "@workspace/common": "workspace:*",
+    "@workspace/domains": "workspace:*",
     "@workspace/typescript-config": "workspace:*",
     "tsx": "^4.7.0",
     "typescript": "^5.7.3"
