--- packages/common/src/domains/index.ts
@@ -1,2 +1,0 @@
-export * from './health'
-export * from './chat'
