--- packages/common/src/index.ts
@@ -1,4 +1,0 @@
-// Export all types from the submodules
-export * from './domains'
-
-export * from './types'
