--- packages/common/src/types/enums.ts
@@ -1,8 +1,0 @@
-/**
- * Defines actors in a chat conversation.
- */
-export enum ChatActor {
-  USER = 'user',
-  ASSISTANT = 'assistant',
-  SYSTEM = 'system',
-}
