--- packages/common/src/types/index.ts
@@ -1,4 +1,0 @@
-// filepath: /Users/perlinde/repositories/digital-office/packages/common/src/types/index.ts
-export * from './id'
-export * from './enums'
-
