+++ packages/domains/src/chat/schema.ts
