+++ packages/domains/src/health/schema.ts
