+++ packages/domains/src/health/use-cases.ts
