+++ packages/types/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@workspace/api",
+  "name": "@workspace/types",
   "version": "0.0.0",
   "private": true,
   "main": "dist/index.js",
@@ -7,8 +7,7 @@
     "access": "public"
   },
   "exports": {
-    "./domains": "./src/domains/index.ts",
-    "./types": "./src/types/index.ts"
+    ".": "./src/index.ts"
   },
   "scripts": {
     "build": "tsc",
@@ -23,7 +22,6 @@
     "zod": "^3.24.2"
   },
   "devDependencies": {
-    "@workspace/typescript-config": "workspace:*",
-    "@workspace/domains": "workspace:*"
+    "@workspace/typescript-config": "workspace:*"
   }
 }
